--- src/lbpDashboard/sparkChart/spark-chart.html
@@ -1,5 +1,0 @@
-<template>
-  <div ref="container" class="sparkChartContainer">
-    <div ref="sparkChart" class="chart"></div>
-  </div>
-</template>
