+++ src/resources/elements/sparkChart/spark-chart.html
@@ -1,0 +1,7 @@
+<template>
+  <div ref="container" class="sparkChartContainer">
+    <div ref="sparkChart" class="chart">
+      <div if.bind="loading" class="isLoading"><i class="spinner fas fa-circle-notch fa-spin" aria-hidden="true"></i></div>
+    </div>
+  </div>
+</template>
